Move ctrlOrCmdPressed
